# chore: Implement AddMany method