--- a/src/core/string_set_test.cc
+++ b/src/core/string_set_test.cc
@@ -552,4 +552,35 @@ void BM_Add(benchmark::State& state) {
 }
 BENCHMARK(BM_Add);
 
+void BM_AddMany(benchmark::State& state) {
+  vector<string> strs;
+  mt19937 generator(0);
+  StringSet ss;
+  unsigned elems = 100000;
+  for (size_t i = 0; i < elems; ++i) {
+    string str = random_string(generator, 16);
+    strs.push_back(str);
+  }
+  ss.Reserve(elems);
+  array<string_view, 32> str_views;
+
+  while (state.KeepRunning()) {
+    unsigned offset = 0;
+    while (offset < elems) {
+      bool res[str_views.size()];
+      unsigned len = min(elems - offset, 32u);
+      for (size_t i = 0; i < len; ++i) {
+        str_views[i] = strs[offset + i];
+      }
+      offset += len;
+      ss.AddMany(str_views.data(), len, UINT32_MAX, res);
+    }
+    state.PauseTiming();
+    ss.Clear();
+    ss.Reserve(elems);
+    state.ResumeTiming();
+  }
+}
+BENCHMARK(BM_AddMany);
+
 }  // namespace dfly