@kostasrim 2024-10-04
@romange 

```
2024-10-04T14:44:23.4115778Z 60: F20241004 14:44:23.410632 22120 init.cc:25] [../src/core/dense_set.h:345]: assert(capacity_log_ > 0) failed!
2024-10-04T14:44:23.4117519Z 60: *** Check failure stack trace: ***
2024-10-04T14:44:23.4154314Z 60:     @          0x116db03  google::LogMessage::Fail()
2024-10-04T14:44:23.4184591Z 60:     @          0x116da4a  google::LogMessage::SendToLog()
2024-10-04T14:44:23.4193535Z 60: F20241004 14:44:23.418989 22119 init.cc:25] [../src/core/dense_set.h:345]: assert(capacity_log_ > 0) failed!
2024-10-04T14:44:23.4194758Z 60: *** Check failure stack trace: ***
2024-10-04T14:44:23.4225910Z 60:     @          0x116db03  google::LogMessage::Fail()
2024-10-04T14:44:23.4242747Z 60:     @          0x116d23f  google::LogMessage::Flush()
2024-10-04T14:44:23.4252989Z 60:     @          0x116da4a  google::LogMessage::SendToLog()
2024-10-04T14:44:23.4292718Z 60:     @          0x1170f56  google::LogMessageFatal::~LogMessageFatal()
2024-10-04T14:44:23.4310424Z 60:     @          0x116d23f  google::LogMessage::Flush()
2024-10-04T14:44:23.4322775Z 60:     @          0x116474d  __assert_fail
2024-10-04T14:44:23.4358823Z 60:     @           0x77537c  dfly::DenseSet::BucketId()
2024-10-04T14:44:23.4361239Z 60:     @          0x1170f56  google::LogMessageFatal::~LogMessageFatal()
2024-10-04T14:44:23.4389632Z 60:     @          0x116474d  __assert_fail
2024-10-04T14:44:23.4393117Z 60:     @           0xcfef51  dfly::DenseSet::Prefetch()
2024-10-04T14:44:23.4424961Z 60:     @           0x77537c  dfly::DenseSet::BucketId()
2024-10-04T14:44:23.4454083Z 60:     @           0xd04e51  dfly::StringSet::AddMany()
2024-10-04T14:44:23.4460588Z 60:     @           0xcfef51  dfly::DenseSet::Prefetch()
2024-10-04T14:44:23.4461702Z 60:     @           0x75cbf9  dfly::(anonymous namespace)::StringSetWrapper::Add()
2024-10-04T14:44:23.4467360Z 60:     @           0x75fd58  dfly::(anonymous namespace)::OpAdd()
```